chore(hooks): hoist useIsMobile to @/hooks (shared; M5 mobile-pass prep)
Composition (M5a/M5b mobile pass) + knowledge now share one viewport hook. The
knowledge module re-exports from @/hooks/useIsMobile so its existing importers
(KnowledgePage, JobsPage) are unchanged.

VERIFY: KnowledgePage.test.tsx 4/4 green (re-export transparent).
